Add wait conditions
With the move to OS::Nova::Server, stacks go complete immediately
irrespective of whether the stack actually builds successfully.  This
change adds to wait conditions: the first waiting for the 4 other nodes
to build before the primary controller (where ansible runs) is built,
and the second waits for the primary controller to successfully build
before the stack is marked complete.

We also remove the apt_common_repos override since we now default to a
more sane value.