Add trailing newline to /etc/ssh/sshd_config
Commit 1dae4cc causes an issue on public cloud as our Trusty image has
no trailing newline in /etc/ssh/sshd_config.  If ansible tries to
append a line (which it seems to do w/ MaxSessions), it ends up writing
it on the same line as the last line in the file
(PasswordAuthentication yes).  This change simply writes a newline to
the end of the file before ansible runs.