Add run_ansible parameter and fix raw_url variable
This commit adds a new heat parameter, run_ansible, which determines
if ansible should run after the stack is created.  In CI systems for
example, we may want the actual ansible run to happen in a separate
task run over ssh.

Also, we fix a bug where raw_url wasn't getting created correctly if
the git URL had a '.git' on the end.  This change simply removes the
.git before transforming the URL into the raw URL.