Remove dependent wait handle
Currently, we often see node1 not getting built even though the other
nodes come up and curl swift successfully.  This change, with
@miguelgrinberg, creates 5 individual signals / signal handlers.  We
now also loop an ansible ping (timing out after 5 mins) to verify that
all nodes are up before doing the ansible run.

Secondly, the code that actually does the ansible run is now saved to a
file and that file gets executed.  The reason for this is to allow you
to a) run the appropriate playbooks even if you set run_ansible to
False and b) allows jenkins to run the ansible run in a seperate job
over ssh.

Lastly, we commit a few scripts which are used as wrappers by jenkins.