Add Ceph support
This commit adds support for deploying a Ceph cluster and enabling it
for nova, cinder and glance.

The variables DEPLOY_CEPH and CEPH_NODE_COUNT have been introduced. When
DEPLOY_CEPH == 'yes' additional nodes are created as OSD hosts. The
number is determined by CEPH_NODE_COUNT and defaults to 1, with a
maximum value of 10,  when DEPLOY_CEPH == 'yes' and 0 otherwise. Each
OSD host gets created with one journal disk and five OSD disks.

controller_primary.sh has been modified to use the deploy.sh script in
rpc-openstack instead of run-playbooks.sh in os-ansible-deployment. This
was done because the ceph playbook needs to be run partway through the
run-playbook.sh script and the deploy.sh script has already done the
work to break this out.