Change default flavour used to build instances
performance1-8 is being phased out. There is currently limited spare
capacity of this flavour. This commit changes the default flavour to
general1-8.

general1-8 does not have an ephemeral disk. This is used for deploying
cinder and swift. This commit uses Cloud Block Storage to attach an
additional disk to the instances to replace the ephemeral disk lost by
switching to the general1-8 flavour.

Closes-Issue: https://github.com/rcbops/rpc-heat/issues/5