Force neutron_l2_population to True
In liberty and mitaka, openstack-ansible defaults neutron_l2_population
to False and enables it when l3ha is enabled (which happens when you
have > 1 neutron agents containers).

We will need to investigate to see how we can use l3ha on our
multi-node deploys, but for now we enable neutron_l2_population so our
instances attached to a tenant network are accessible.