Create new dependency using SoftwareDeployment
The scripts when the instances startup require the volumes to already be
attached. This means there is a race between the scripts and the volume
attachment resource.

This commit separates the user_data scripts from the instances and uses
The SoftwareConfig and SoftwareDeployment resources to ensure volume
attachment is completed prior to script execution.

This commit adds a submodule to pull in the OpenStack heat-templates
repo so that the instances can be configured for use with the
SoftwareDeployment resource.